leverage ncatelli/isa-mos6502 (#100)
* include isa-mos6502 and bump spasm

* Remove mnemonic module

* remove mnemonic references

* Replace references to AddressingMode with isa_mos6502

* Move AddressingModeType to isa lib

* Start implementing variant conversion

* Use InstructionVariant as emitter

* Bump spasm version in Cargo.toml

* Remove StaticInstruction in favor of InstructionVariant

* Fix error handling of instructionvariant

* Pin isa-mos6502 to v0.2.0 tag

* fix clippy lints

+++ src/lib.rs
@@ -93,7 +93,7 @@ impl Emitter<Vec<u8>> for Vec<Origin<Vec<u8>>> {
             .map(|(bytecode, pad_size)| {
                 bytecode
                     .into_iter()
-                    .chain(vec![0 as u8].into_iter().cycle().take(pad_size))
+                    .chain(vec![0u8].into_iter().cycle().take(pad_size))
                     .collect::<Vec<u8>>()
             })
             .flatten()
